Add files via upload
This Python mini project analyzes a given number and provides useful insights such as whether the number is positive, negative, or zero, checks if it is even or odd, and determines whether it is a single-digit or multi-digit number. The project is designed for beginners to practice conditional statements, input handling, and basic arithmetic operations in Python.